Factor out safecall parsing
