+++ src/plugins/lightscript.js
@@ -551,56 +551,6 @@ pp.isBitwiseOp = function () {
   );
 };
 
-pp.parseSafeCall = function(expr, startPos, startLoc) {
-  this.expect(tt.parenL);
-  const node = this.startNodeAt(startPos, startLoc);
-  node.callee = expr;
-  node.arguments = this.parseCallExpressionArguments(tt.parenR, false);
-  node.safe = true;
-  return this.finishNode(node, "CallExpression");
-};
-
-pp.parseExistential = function(expr, startPos, startLoc) {
-  const node = this.startNodeAt(startPos, startLoc);
-  node.argument = expr;
-  return this.finishNode(node, "ExistentialExpression");
-};
-
-pp.parseQuestionSubscript = function(lhs, startPos, startLoc, noCalls) {
-  const questionPos = this.state.pos;
-  const questionLine = this.state.curLine;
-  const state = this.state.clone();
-
-  this.eat(tt.question);
-
-  // `?(` = safecall or poorly-formatted ternary
-  // TODO: lint rule for ternaries recommending space after ?
-  if (!noCalls && this.match(tt.parenL) && this.state.pos === (questionPos + 1)) {
-    try {
-      return this.parseSafeCall(lhs, startPos, startLoc);
-    } catch (e) {
-      this.state = state;
-      this.eat(tt.question);
-    }
-  }
-
-  // If the next token startsExpr, this is a ternary -- unwind recursive descent
-  if (this.state.type.startsExpr && this.state.curLine === questionLine) {
-    this.state = state;
-    return null;
-  }
-
-  // Otherwise this is an existential
-  return this.parseExistential(lhs, startPos, startLoc);
-};
-
-// Convert an existential to an optional flow parameter.
-// Resolves grammar ambiguity in arrow function arg lists.
-pp.existentialToParameter = function(node) {
-  node.argument.optional = true;
-  return node.argument;
-};
-
 pp.parseMatchExpression = function (node) {
   return this.parseMatch(node, true);
 };
@@ -708,7 +658,6 @@ pp.parseMatchCaseBinding = function (node) {
   }
   node.binding = this.parseBindingAtom();
 };
-
 
 export default function (instance) {
