+++ src/registerPlugins.js
@@ -2,6 +2,7 @@ import lightscriptPlugin from "./plugins/lightscript";
 import estreePlugin from "./plugins/estree";
 import flowPlugin from "./plugins/flow";
 import jsxPlugin from "./plugins/jsx";
+import safeCallExistentialPlugin from "./plugins/safeCallExistential";
 
 function noncePlugin() {}
 
@@ -10,6 +11,8 @@ export default function registerPlugins(plugins) {
   plugins.estree = estreePlugin;
   plugins.flow = flowPlugin;
   plugins.jsx = jsxPlugin;
+  plugins.safeCallExpression = safeCallExistentialPlugin;
+  plugins.existentialExpression = safeCallExistentialPlugin;
 
   // Plugins with string tags only
   ([
